aimbot + reworked esp

+++ src/esp.py
@@ -56,7 +56,7 @@ def esp():
                 elif entity_team_id == 3:  # ct
                     
                     pm.write_float(glow_manager + entity_glow * 0x38 + 0x8, float(0))   # R
-                    pm.write_float(glow_manager + entity_glow * 0x38 + 0xC, float(0))   # G
-                    pm.write_float(glow_manager + entity_glow * 0x38 + 0x10, float(1))  # B
+                    pm.write_float(glow_manager + entity_glow * 0x38 + 0xC, float(1))   # G
+                    pm.write_float(glow_manager + entity_glow * 0x38 + 0x10, float(0))  # B
                     pm.write_float(glow_manager + entity_glow * 0x38 + 0x14, float(1))  # Alpha
                     pm.write_int(glow_manager + entity_glow * 0x38 + 0x28, 1)           # Enable glow
